Moves /hash/sha3 to dh/sidh/internal/shake
