+++ dh/sidh/internal/shake/keccakf.go
@@ -4,7 +4,7 @@
 
 //  +build !amd64 appengine gccgo
 
-package sha3
+package shake
 
 // rc stores the round constants for use in the ι step.
 var rc = [24]uint64{
