+++ dh/sidh/internal/shake/testdata/keccakKats.json.deflate
