+++ dh/sidh/internal/shake/xor_unaligned.go
@@ -5,7 +5,7 @@
 // +build amd64 386 ppc64le
 // +build !appengine
 
-package sha3
+package shake
 
 import "unsafe"
